Docker/Multi-service_Docker/web/requirements.txt
fastapi==0.101.0
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
httpx==0.25.0
psycopg2-binary==2.9.7
requests==2.31.0
